+++ include/geos/geom/LineString.h
@@ -103,7 +103,7 @@ public:
     int getBoundaryDimension() const override;
 
     /// Returns coordinate dimension.
-    int getCoordinateDimension() const override;
+    uint8_t getCoordinateDimension() const override;
 
     /**
      * \brief
